pyproject.toml
[project]
name = "pygullian-checker"
version = "0.0.1"
authors = [
  { name="Marcel Guinhos de Menezes", email="mguinhos@gmail.com" },
]
description = "A type checker for the gullian language"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pygullian-parser @ git+https://github.com/GullianLanguage/pygullian-parser",
]

[project.urls]
"Homepage" = "https://github.com/GullianLanguage/pygullian-checker"
"Bug Tracker" = "https://github.com/GullianLanguage/pygullian-checker/issues"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true